Allow console output to be visible for command line users
If you use Console.Out.WriteLine with a Windows Forms application straightforwardly, the output will not appear in a terminal. It will only appear in the debugger if attached.

This change adds a directive to attach to the console, so that the output appears. It sends a key to clear and detaches from the console when it's done.

Because attaching to consoles is not  exposed through Windows Forms this uses PInvoke with Win32.